Add registration of adapter endpoints for all standard profiles, putting HomeAutomation in endpoint 2.
The deviates from koenkk/zigbee-herdsman, which places Home Automation in endpoint 1, it's unknown if
their approach results in a higher compatibility or not, it might just be that was the first implemented
in zigbee-herdsman.

Some zigbee devices will do the right thing and match the endpoint, some wont. We do know certain devices
need additional endpoints which are specific to their implementation. Though we're currently not implementing
them.